+++ lib/screens/chat_screen.dart
@@ -317,18 +317,6 @@ class ChatScreen extends HookConsumerWidget {
                 color: colors.backgroundPrimary,
                 height: inputAreaHeight.value + safeAreaBottom + 48.h,
               ),
-              if (noticeMessage.value != null)
-                Positioned(
-                  top: safeAreaTop,
-                  left: 0,
-                  right: 0,
-                  child: WnSystemNotice(
-                    key: ValueKey(noticeMessage.value),
-                    title: noticeMessage.value!,
-                    type: WnSystemNoticeType.error,
-                    onDismiss: dismissNotice,
-                  ),
-                ),
               SafeArea(
                 bottom: false,
                 child: Column(
@@ -363,6 +351,14 @@ class ChatScreen extends HookConsumerWidget {
                               )
                             : null,
                       ),
+                      systemNotice: noticeMessage.value != null
+                          ? WnSystemNotice(
+                              key: ValueKey(noticeMessage.value),
+                              title: noticeMessage.value!,
+                              type: WnSystemNoticeType.error,
+                              onDismiss: dismissNotice,
+                            )
+                          : null,
                     ),
                     if (isSearchActive.value) ...[
                       Padding(